// lib/types/member/member.ts
import { MemberAuthType, MemberStatus, MemberType } from "@/lib/enums/member";
import { MeFollowed, MeLiked, TotalCounter } from "../common";

export interface Member {
  _id: string;
  memberType: MemberType;
  memberStatus: MemberStatus;
  memberAuthType: MemberAuthType;
  memberPhone: string;
  memberNick: string;
  memberPassword?: string;
  memberFullName?: string;
  memberImage?: string;
  memberAddress?: string;
  memberDesc?: string;
  memberProperties: number;
  memberRank: number;
  memberArticles: number;
  memberPoints: number;
  memberLikes: number;
  memberFollowers?: number;
  memberFollowings?: number;
  memberViews: number;
  memberComments: number;
  memberWarnings: number;
  memberBlocks: number;
  deletedAt?: Date;
  createdAt: Date;
  updatedAt: Date;
  // Enable for authentications
  meLiked?: MeLiked[];
  meFollowed?: MeFollowed[];
  accessToken?: string;
}

export interface Members {
  list: Member[];
  metaCounter: TotalCounter[];
}